trying to upload my working pose generator
